+++ CesiumGeospatial/src/LocalHorizontalCoordinateSystem.cpp
@@ -77,34 +77,37 @@ LocalHorizontalCoordinateSystem::LocalHorizontalCoordinateSystem(
       GlobeTransforms::eastNorthUpToFixedFrame(originEcef, ellipsoid);
 
   // Construct a matrix to swap and invert axes as necessary
-  glm::dmat4 localToEnu(
-      glm::dvec4(directionToEnuVector(xAxisDirection), 0.0),
-      glm::dvec4(directionToEnuVector(yAxisDirection), 0.0),
-      glm::dvec4(directionToEnuVector(zAxisDirection), 0.0),
-      glm::dvec4(0.0, 0.0, 0.0, 1.0));
+  glm::dmat3 localToEnuAndScale(
+      scaleToMeters * directionToEnuVector(xAxisDirection),
+      scaleToMeters * directionToEnuVector(yAxisDirection),
+      scaleToMeters * directionToEnuVector(zAxisDirection));
 
-  glm::dmat4 scale{glm::dmat3(scaleToMeters)};
-
-  this->_localToEcef = enuToFixed * localToEnu * scale;
+  this->_localToEcef = enuToFixed * glm::dmat4(localToEnuAndScale);
   this->_ecefToLocal = glm::affineInverse(this->_localToEcef);
 }
 
 glm::dvec3 LocalHorizontalCoordinateSystem::localPositionToEcef(
-    const glm::dvec3& localPosition) const {
+    const glm::dvec3& localPosition) const noexcept {
   return glm::dvec3(this->_localToEcef * glm::dvec4(localPosition, 1.0));
 }
 
 glm::dvec3 LocalHorizontalCoordinateSystem::ecefPositionToLocal(
-    const glm::dvec3& ecefPosition) const {
+    const glm::dvec3& ecefPosition) const noexcept {
   return glm::dvec3(this->_ecefToLocal * glm::dvec4(ecefPosition, 1.0));
 }
 
 glm::dvec3 LocalHorizontalCoordinateSystem::localDirectionToEcef(
-    const glm::dvec3& localDirection) const {
+    const glm::dvec3& localDirection) const noexcept {
   return glm::dvec3(this->_localToEcef * glm::dvec4(localDirection, 0.0));
 }
 
 glm::dvec3 LocalHorizontalCoordinateSystem::ecefDirectionToLocal(
-    const glm::dvec3& ecefDirection) const {
+    const glm::dvec3& ecefDirection) const noexcept {
   return glm::dvec3(this->_ecefToLocal * glm::dvec4(ecefDirection, 0.0));
+}
+
+glm::dmat4 LocalHorizontalCoordinateSystem::computeTransformationToAnotherLocal(
+    const LocalHorizontalCoordinateSystem& target) const noexcept {
+  return target.getEcefToLocalTransformation() *
+         this->getLocalToEcefTransformation();
 }