+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### ? - ?
+
+##### Additions :tada:
+
+- Added support for the `KHR_materials_variants` extension to the glTF reader and writer.
+
 ### v0.24.0 - 2023-05-01
 
 ##### Additions :tada: