+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### ? - ?
+
+##### Additions :tada:
+
+- Added `computeTransformationToAnotherLocal` method to `LocalHorizontalCoordinateSystem`.
+
 ### v0.24.0 - 2023-05-01
 
 ##### Additions :tada: