+++ CHANGES.md
@@ -5,6 +5,7 @@
 ##### Additions :tada:
 
 - Added `computeTransformationToAnotherLocal` method to `LocalHorizontalCoordinateSystem`.
+- Added support for the `KHR_materials_variants` extension to the glTF reader and writer.
 
 ### v0.24.0 - 2023-05-01