+++ CHANGES.md
@@ -7,7 +7,9 @@
 - Added support for the `KHR_materials_variants` extension to the glTF reader and writer.
 
 ##### Fixes :wrench:
-- Upon Tileset Load Failure, cesium-native will always log warning/error messages even if the failure callback is set. 
+
+- On Tileset Load Failure, warning/error messages will always be logged even if the failure callback is set. 
+- Fixed a bug that caused meshes to be missing entirely when upsampled from a parent with `UNSIGNED_BYTE` indices.
 
 ### v0.24.0 - 2023-05-01