+++ CHANGES.md
@@ -4,11 +4,13 @@
 
 ##### Additions :tada:
 
+- Added `computeTransformationToAnotherLocal` method to `LocalHorizontalCoordinateSystem`.
 - Added support for the `KHR_materials_variants` extension to the glTF reader and writer.
 - The Caching Asset Accessor will now automatically unzip data if gzipped data is detected.
 
 ##### Fixes :wrench:
 
+- On Tileset Load Failure, warning/error messages will always be logged even if the failure callback is set. 
 - Fixed a bug that caused meshes to be missing entirely when upsampled from a parent with `UNSIGNED_BYTE` indices.
 
 ### v0.24.0 - 2023-05-01